Align email date format with event date format

+++ front/app/components/admin/Email/DraftCampaignRow.tsx
@@ -61,7 +61,7 @@ const DraftCampaignRow = ({ campaign, context }: Props) => {
                 <FormattedDate
                   value={campaign.attributes.scheduled_at}
                   timeZone={timeZone}
-                  dateStyle="short"
+                  dateStyle="long"
                   timeStyle="short"
                 />
               </Text>
