+++ front/app/containers/Admin/messaging/CustomEmails/Show/index.tsx
@@ -246,7 +246,7 @@ const Show = () => {
                   <FormattedDate
                     value={campaign.data.attributes.scheduled_at}
                     timeZone={timeZone}
-                    dateStyle="short"
+                    dateStyle="long"
                     timeStyle="short"
                   />
                 </Text>
