+++ front/app/containers/Admin/messaging/Edit/index.tsx
@@ -100,7 +100,7 @@ const Edit = ({ campaignType }: EditProps) => {
                 <FormattedDate
                   value={campaign.data.attributes.scheduled_at}
                   timeZone={timeZone}
-                  dateStyle="short"
+                  dateStyle="long"
                   timeStyle="short"
                 />
               </Text>
